feat: add a notice class and before_notify hook
The main goal of this PR was to add a `before_notify` hook as part of the configuration. This allows the library to be configured to intercept, inspect, and alter a error notice before it is sent off to Honeybadger.

To make this easier, a new `Notice` class has been added so that there is a single object that gets passed into the hook. The payload is generate upon instantiation and can be later inspected or altered in the hook.

This preemptive generation isn't ideal because it's a little cumbersome to work with the notice payload. However, the plugins implement the `generate_payload` method which takes in the formatted notice payload. After this, I'd like to take a stab at making it easier to manipulate a notice without having to dig into the final payload.