Shutdown on SIGTERM
Since lambda uses TERM for the shutdown signal, and atexit does not
cover the term signal, this change listens for TERM and does a proper
exit (which then uses atexit).

It sounds like there is not really a perfect way to handle this, but if
there are any other signal callbacks, then we assume that some other
listener will handle a proper shutdown and delegate to that.